+++ common/src/main/java/com/kwwsyk/endinv/common/options/IConfigValue.java
@@ -21,4 +21,19 @@ public interface IConfigValue<T> {
             }
         };
     }
+
+    static <T> IConfigValue<T> of(Supplier<T> getter, Consumer<T> setter, Runnable onset){
+        return new IConfigValue<T>() {
+            @Override
+            public T get() {
+                return getter.get();
+            }
+
+            @Override
+            public void set(T t) {
+                setter.accept(t);
+                onset.run();
+            }
+        };
+    }
 }