+++ fabric/src/main/java/com/kwwsyk/endinv/fabric/event/Commands.java
@@ -1,5 +1,6 @@
 package com.kwwsyk.endinv.fabric.event;
 
+import com.kwwsyk.endinv.common.commands.ConfigCommand;
 import com.kwwsyk.endinv.common.commands.EndInvCommand;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 
@@ -9,6 +10,9 @@ public final class Commands {
     }
 
     public static void register() {
-        CommandRegistrationCallback.EVENT.register((dispatcher, registryAccess, environment) -> EndInvCommand.register(dispatcher));
+        CommandRegistrationCallback.EVENT.register((dispatcher, registryAccess, environment) -> {
+            EndInvCommand.register(dispatcher);
+            ConfigCommand.register(dispatcher);
+        });
     }
 }