+++ fabric/src/main/java/com/kwwsyk/endinv/fabric/event/FabricEvents.java
@@ -9,6 +9,7 @@ public final class FabricEvents {
         Commands.register();
         LevelEvents.register();
         PlayerEvents.register();
-        LootEvent.register();
+        //LootEvent.register();
+        BlockBreakRedirect.register();
     }
 }