+++ forge/src/main/java/com/kwwsyk/endinv/forge/events/CommandsReg.java
@@ -1,6 +1,7 @@
 package com.kwwsyk.endinv.forge.events;
 
 import com.kwwsyk.endinv.common.ModInfo;
+import com.kwwsyk.endinv.common.commands.ConfigCommand;
 import com.kwwsyk.endinv.common.commands.EndInvCommand;
 import net.minecraftforge.event.RegisterCommandsEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -12,5 +13,6 @@ public class CommandsReg {
     @SubscribeEvent
     public static void regCommands(final RegisterCommandsEvent event){
         EndInvCommand.register(event.getDispatcher());
+        ConfigCommand.register(event.getDispatcher());
     }
 }